services/ai-service/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
redis==4.5.5
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tensorflow==2.12.0
torch==2.0.1
transformers==4.30.2
gunicorn==20.1.0